create a reference file
Since the main working script is a live script. I have made a reference script that will not work without commenting out sections of the code. This script is meant to serve as as easy way to view changes in the code on GitHub without having to download the main(live) script.