resit update 5 Task 1.3 complete & Task 1.5 begins
The plots for the shock angles over and under the wedge are now complete. I may decide to make them more accurate by calculating the horizontal and vertical components of the Mach number using the obliques shock angle instead of setting an x-componnt of arbitrary length, but that will be later in future.

Task 1.5 has now begun. And I'm not sure why I keep getting complex numbers whenever I try calculating the 150 psi working condition, but we have time to fix that so it should be alright.